requirements.txt
requests
pandas
google-cloud-storage
google-cloud-bigquery
pyarrow
db-dtypes
fsspec
pytz